Updating Rhino.DocObjects.LeaderObject
Mostly a test to see if pushing code to Rhinocommon will work this way.
Secondarily is testing to see if the code I've updated is correct and
useful.  In this case it is updating the Rhino.DocObjects.LeaderObject
to access its Rhino.Geometry.Leader so that users can .CommitChanges()
after editing the leader.

Is this code correct?

Am I correct in assuming that I could contribute to the code in this
manner, as well as potentially adding examples to the documentation and
adding descriptions where it is missing in the documentation?